# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a test task for Erply using Erply API.",
        "title": "Erply Test Task",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/getCustomerByID": {
            "post": {
                "description": "Retrieves customer information either from local database (if customer data in database is less than 10 minutes old) or by making a request to ERPLY API",
                "produces": [
                    "application/json"
                ],
                "summary": "Get customer information by customer ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client code",
                        "name": "clientCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session key",
                        "name": "sessionKey",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CustomerRecord"
                        }
                    }
                }
            }
        },
        "/getCustomers": {
            "post": {
                "description": "Retrieves customer information by making a request to ERPLY API",
                "produces": [
                    "application/json"
                ],
                "summary": "Get customer information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client code",
                        "name": "clientCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session key",
                        "name": "sessionKey",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CustomerResponse"
                        }
                    }
                }
            }
        },
        "/getSessionKeyInfo": {
            "post": {
                "description": "Retrieves session key information by making a request to ERPLY API",
                "produces": [
                    "application/json"
                ],
                "summary": "Get session key information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client code",
                        "name": "clientCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session key",
                        "name": "sessionKey",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetSessionKeyInfoResponse"
                        }
                    }
                }
            }
        },
        "/saveCustomer": {
            "post": {
                "description": "Saves customer information by making a request to ERPLY API",
                "produces": [
                    "application/json"
                ],
                "summary": "Save customer information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client code",
                        "name": "clientCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session key",
                        "name": "sessionKey",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Full name",
                        "name": "fullName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SaveCustomerResponse"
                        }
                    }
                }
            }
        },
        "/verifyUser": {
            "post": {
                "description": "Verifies the user credentials by making a request to ERPLY API",
                "produces": [
                    "application/json"
                ],
                "summary": "Verify user credentials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client code",
                        "name": "clientCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CustomerRecord": {
            "type": "object",
            "properties": {
                "EDI": {
                    "type": "string"
                },
                "GLN": {
                    "type": "string"
                },
                "PeppolID": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "address2": {
                    "type": "string"
                },
                "addressTypeID": {
                    "type": "integer"
                },
                "addressTypeName": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "colorStatus": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "companyTypeID": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "countryID": {
                    "type": "string"
                },
                "credit": {
                    "type": "integer"
                },
                "creditCardLastNumbers": {
                    "type": "string"
                },
                "customerBalanceDisabled": {
                    "type": "integer"
                },
                "customerCardNumber": {
                    "type": "string"
                },
                "customerID": {
                    "type": "integer"
                },
                "customerType": {
                    "type": "string"
                },
                "deliveryTypeID": {
                    "type": "integer"
                },
                "doNotSell": {
                    "type": "integer"
                },
                "docuraEDIEnabled": {
                    "type": "integer"
                },
                "eInvoiceEmail": {
                    "type": "string"
                },
                "eInvoiceEnabled": {
                    "type": "integer"
                },
                "eInvoiceReference": {
                    "type": "string"
                },
                "ediType": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emailEnabled": {
                    "type": "integer"
                },
                "emailOptOut": {
                    "type": "integer"
                },
                "euCustomerType": {
                    "type": "string"
                },
                "externalIDs": {
                    "type": "array",
                    "items": {}
                },
                "facebookName": {
                    "type": "string"
                },
                "factoringContractNumber": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "flagStatus": {
                    "type": "integer"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "groupID": {
                    "type": "integer"
                },
                "groupName": {
                    "type": "string"
                },
                "homeStoreID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "integrationCode": {
                    "type": "string"
                },
                "isPOSDefaultCustomer": {
                    "type": "integer"
                },
                "lastModifierEmployeeID": {
                    "type": "integer"
                },
                "lastModifierUsername": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "mailEnabled": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string"
                },
                "operatorIdentifier": {
                    "type": "string"
                },
                "partialTaxExemption": {
                    "type": "integer"
                },
                "payerID": {
                    "type": "integer"
                },
                "paysViaFactoring": {
                    "type": "integer"
                },
                "personTitleID": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "posCouponsDisabled": {
                    "type": "integer"
                },
                "postalCode": {
                    "type": "string"
                },
                "referenceNumber": {
                    "type": "string"
                },
                "rewardPoints": {
                    "type": "integer"
                },
                "rewardPointsDisabled": {
                    "type": "integer"
                },
                "salesBlocked": {
                    "type": "integer"
                },
                "signUpStoreID": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "taxExempt": {
                    "type": "integer"
                },
                "twitterID": {
                    "type": "string"
                }
            }
        },
        "models.CustomerResponse": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CustomerRecord"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.Status"
                }
            }
        },
...
        "models.GetSessionKeyInfoResponse": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SessionKeyRecord"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.Status"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Record"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.Status"
                }
            }
        },
        "models.SaveCustomerRecord": {
            "type": "object",
            "properties": {
                "alreadyExists": {
                    "type": "boolean"
                },
                "clientID": {
                    "type": "integer"
                },
                "customerID": {
                    "type": "integer"
                }
            }
        },
        "models.SaveCustomerResponse": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SaveCustomerRecord"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.SaveCustomerResponseStatus"
                }
            }
        },
        "models.SaveCustomerResponseStatus": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "generationTime": {
                    "type": "number"
                },
                "recordsInResponse": {
                    "type": "integer"
                },
                "recordsTotal": {
                    "type": "integer"
                },
                "request": {
                    "type": "string"
                },
                "requestUnixTime": {
                    "type": "integer"
                },
                "responseStatus": {
                    "type": "string"
                }
            }
        },
        "models.SessionKeyRecord": {
            "type": "object",
            "properties": {
                "creationUnixTime": {
                    "type": "string"
                },
                "expireUnixTime": {
                    "type": "string"
                }
            }
        },
        "models.Status": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "generationTime": {
                    "type": "number"
                },
                "recordsInResponse": {
                    "type": "integer"
                },
                "recordsTotal": {
                    "type": "integer"
                },
                "request": {
                    "type": "string"
                },
                "requestUnixTime": {
                    "type": "integer"
                },
                "responseStatus": {
                    "type": "string"
                }
            }
        }
    }
}